requirements.txt
Django==2.2
django-extra-fields==2.0.2
djangorestframework==3.11.0
markdown==3.1.1
django-filter==2.2.0
djangorestframework-simplejwt==4.4.0
easy-thumbnails==2.7
pytest-django==3.8.0
factory-boy==2.12.0
Pillow==7.0.0
drf-yasg==1.17.0
flake8==3.7.9